+++ src/index.ts
@@ -1,5 +1,5 @@
 
-import Bootstrap5Generator, { Cache, Bootstrap5GeneratorConstructor, BootstrapVariables, BootstrapTheme, BoostrapThemeSCSS } from "./Bootstrap5Generator"
+import Bootstrap5Generator, { Cache, Bootstrap5GeneratorConstructor, BootstrapVariables, BootstrapTheme, BoostrapThemeSCSS, BootstrapComponent, ThemeBackground } from "./Bootstrap5Generator"
 
-export type { Cache, Bootstrap5GeneratorConstructor, BootstrapVariables, BootstrapTheme, BoostrapThemeSCSS }
+export type { Cache, Bootstrap5GeneratorConstructor, BootstrapVariables, BootstrapTheme, BoostrapThemeSCSS, BootstrapComponent, ThemeBackground }
 export { Bootstrap5Generator }