+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Fixed bug on method getHashFromTheme
 
 ## [0.4.0] - 2024-04-17
 - Added color support to theme generation, adding a new color will generate all related classes included the subtle colors (btn-new-color, ...)