+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Support theme variables without "$", "$" will be appended when generating the css
 
 ## [0.6.0] - 2024-05-22
 - Added support to dynamic background